Introduced MinimalisticDataPartChecksums. [#CLICKHOUSE-3660]
