+++ dbms/src/Storages/MergeTree/MergeTreeDataPart.cpp
@@ -30,272 +30,14 @@ namespace DB
 
 namespace ErrorCodes
 {
-    extern const int CHECKSUM_DOESNT_MATCH;
     extern const int FILE_DOESNT_EXIST;
     extern const int NO_FILE_IN_DATA_PART;
     extern const int EXPECTED_END_OF_FILE;
-    extern const int BAD_SIZE_OF_FILE_IN_DATA_PART;
     extern const int CORRUPTED_DATA;
-    extern const int FORMAT_VERSION_TOO_OLD;
-    extern const int UNKNOWN_FORMAT;
-    extern const int UNEXPECTED_FILE_IN_DATA_PART;
     extern const int NOT_FOUND_EXPECTED_DATA_PART;
+    extern const int BAD_SIZE_OF_FILE_IN_DATA_PART;
 }
 
-
-void MergeTreeDataPartChecksum::checkEqual(const MergeTreeDataPartChecksum & rhs, bool have_uncompressed, const String & name) const
-{
-    if (is_compressed && have_uncompressed)
-    {
-        if (!rhs.is_compressed)
-            throw Exception("No uncompressed checksum for file " + name, ErrorCodes::CHECKSUM_DOESNT_MATCH);
-        if (rhs.uncompressed_size != uncompressed_size)
-            throw Exception("Unexpected uncompressed size of file " + name + " in data part", ErrorCodes::BAD_SIZE_OF_FILE_IN_DATA_PART);
-        if (rhs.uncompressed_hash != uncompressed_hash)
-            throw Exception("Checksum mismatch for uncompressed file " + name + " in data part", ErrorCodes::CHECKSUM_DOESNT_MATCH);
-        return;
-    }
-    if (rhs.file_size != file_size)
-        throw Exception("Unexpected size of file " + name + " in data part", ErrorCodes::BAD_SIZE_OF_FILE_IN_DATA_PART);
-    if (rhs.file_hash != file_hash)
-        throw Exception("Checksum mismatch for file " + name + " in data part", ErrorCodes::CHECKSUM_DOESNT_MATCH);
-}
-
-void MergeTreeDataPartChecksum::checkSize(const String & path) const
-{
-    Poco::File file(path);
-    if (!file.exists())
-        throw Exception(path + " doesn't exist", ErrorCodes::FILE_DOESNT_EXIST);
-    UInt64 size = file.getSize();
-    if (size != file_size)
-        throw Exception(path + " has unexpected size: " + toString(size) + " instead of " + toString(file_size),
-            ErrorCodes::BAD_SIZE_OF_FILE_IN_DATA_PART);
-}
-
-
-void MergeTreeDataPartChecksums::checkEqual(const MergeTreeDataPartChecksums & rhs, bool have_uncompressed) const
-{
-    for (const auto & it : rhs.files)
-    {
-        const String & name = it.first;
-
-        if (!files.count(name))
-            throw Exception("Unexpected file " + name + " in data part", ErrorCodes::UNEXPECTED_FILE_IN_DATA_PART);
-    }
-
-    for (const auto & it : files)
-    {
-        const String & name = it.first;
-
-        auto jt = rhs.files.find(name);
-        if (jt == rhs.files.end())
-            throw Exception("No file " + name + " in data part", ErrorCodes::NO_FILE_IN_DATA_PART);
-
-        it.second.checkEqual(jt->second, have_uncompressed, name);
-    }
-}
-
-void MergeTreeDataPartChecksums::checkSizes(const String & path) const
-{
-    for (const auto & it : files)
-    {
-        const String & name = it.first;
-        it.second.checkSize(path + name);
-    }
-}
-
-bool MergeTreeDataPartChecksums::read(ReadBuffer & in)
-{
-    files.clear();
-
-    assertString("checksums format version: ", in);
-    int format_version;
-    readText(format_version, in);
-    assertChar('\n', in);
-
-    switch (format_version)
-    {
-        case 1:
-            return false;
-        case 2:
-            return read_v2(in);
-        case 3:
-            return read_v3(in);
-        case 4:
-            return read_v4(in);
-        default:
-            throw Exception("Bad checksums format version: " + DB::toString(format_version), ErrorCodes::UNKNOWN_FORMAT);
-    }
-}
-
-bool MergeTreeDataPartChecksums::read_v2(ReadBuffer & in)
-{
-    size_t count;
-
-    readText(count, in);
-    assertString(" files:\n", in);
-
-    for (size_t i = 0; i < count; ++i)
-    {
-        String name;
-        Checksum sum;
-
-        readString(name, in);
-        assertString("\n\tsize: ", in);
-        readText(sum.file_size, in);
-        assertString("\n\thash: ", in);
-        readText(sum.file_hash.first, in);
-        assertString(" ", in);
-        readText(sum.file_hash.second, in);
-        assertString("\n\tcompressed: ", in);
-        readText(sum.is_compressed, in);
-        if (sum.is_compressed)
-        {
-            assertString("\n\tuncompressed size: ", in);
-            readText(sum.uncompressed_size, in);
-            assertString("\n\tuncompressed hash: ", in);
-            readText(sum.uncompressed_hash.first, in);
-            assertString(" ", in);
-            readText(sum.uncompressed_hash.second, in);
-        }
-        assertChar('\n', in);
-
-        files.insert(std::make_pair(name, sum));
-    }
-
-    return true;
-}
-
-bool MergeTreeDataPartChecksums::read_v3(ReadBuffer & in)
-{
-    size_t count;
-
-    readVarUInt(count, in);
-
-    for (size_t i = 0; i < count; ++i)
-    {
-        String name;
-        Checksum sum;
-
-        readBinary(name, in);
-        readVarUInt(sum.file_size, in);
-        readPODBinary(sum.file_hash, in);
-        readBinary(sum.is_compressed, in);
-
-        if (sum.is_compressed)
-        {
-            readVarUInt(sum.uncompressed_size, in);
-            readPODBinary(sum.uncompressed_hash, in);
-        }
-
-        files.emplace(std::move(name), sum);
-    }
-
-    return true;
-}
-
-bool MergeTreeDataPartChecksums::read_v4(ReadBuffer & from)
-{
-    CompressedReadBuffer in{from};
-    return read_v3(in);
-}
-
-void MergeTreeDataPartChecksums::write(WriteBuffer & to) const
-{
-    writeString("checksums format version: 4\n", to);
-
-    CompressedWriteBuffer out{to, CompressionSettings(CompressionMethod::LZ4), 1 << 16};
-    writeVarUInt(files.size(), out);
-
-    for (const auto & it : files)
-    {
-        const String & name = it.first;
-        const Checksum & sum = it.second;
-
-        writeBinary(name, out);
-        writeVarUInt(sum.file_size, out);
-        writePODBinary(sum.file_hash, out);
-        writeBinary(sum.is_compressed, out);
-
-        if (sum.is_compressed)
-        {
-            writeVarUInt(sum.uncompressed_size, out);
-            writePODBinary(sum.uncompressed_hash, out);
-        }
-    }
-}
-
-void MergeTreeDataPartChecksums::addFile(const String & file_name, UInt64 file_size, MergeTreeDataPartChecksum::uint128 file_hash)
-{
-    files[file_name] = Checksum(file_size, file_hash);
-}
-
-void MergeTreeDataPartChecksums::add(MergeTreeDataPartChecksums && rhs_checksums)
-{
-    for (auto & checksum : rhs_checksums.files)
-        files[std::move(checksum.first)] = std::move(checksum.second);
-
-    rhs_checksums.files.clear();
-}
-
-/// Checksum computed from the set of control sums of .bin files.
-void MergeTreeDataPartChecksums::summaryDataChecksum(SipHash & hash) const
-{
-    /// We use fact that iteration is in deterministic (lexicographical) order.
-    for (const auto & it : files)
-    {
-        const String & name = it.first;
-        const Checksum & sum = it.second;
-
-        if (!endsWith(name, ".bin"))
-            continue;
-
-        UInt64 len = name.size();
-        hash.update(len);
-        hash.update(name.data(), len);
-        hash.update(sum.uncompressed_size);
-        hash.update(sum.uncompressed_hash);
-    }
-}
-
-String MergeTreeDataPartChecksums::toString() const
-{
-    WriteBufferFromOwnString out;
-    write(out);
-    return out.str();
-}
-
-MergeTreeDataPartChecksums MergeTreeDataPartChecksums::parse(const String & s)
-{
-    ReadBufferFromString in(s);
-    MergeTreeDataPartChecksums res;
-    if (!res.read(in))
-        throw Exception("Checksums format is too old", ErrorCodes::FORMAT_VERSION_TOO_OLD);
-    assertEOF(in);
-    return res;
-}
-
-const MergeTreeDataPartChecksums::Checksum * MergeTreeDataPart::tryGetChecksum(const String & name, const String & ext) const
-{
-    if (checksums.empty())
-        return nullptr;
-
-    const auto & files = checksums.files;
-    const auto file_name = escapeForFileName(name) + ext;
-    auto it = files.find(file_name);
-
-    return (it == files.end()) ? nullptr : &it->second;
-}
-
-const MergeTreeDataPartChecksums::Checksum * MergeTreeDataPart::tryGetBinChecksum(const String & name) const
-{
-    return tryGetChecksum(name, ".bin");
-}
-
-const MergeTreeDataPartChecksums::Checksum * MergeTreeDataPart::tryGetMrkChecksum(const String & name) const
-{
-    return tryGetChecksum(name, ".mrk");
-}
 
 static ReadBufferFromFile openForReading(const String & path)
 {
@@ -925,6 +667,29 @@ bool MergeTreeDataPart::hasColumnFiles(const String & column) const
     String escaped_column = escapeForFileName(column);
     return Poco::File(prefix + escaped_column + ".bin").exists()
         && Poco::File(prefix + escaped_column + ".mrk").exists();
+}
+
+
+const MergeTreeDataPartChecksums::Checksum * MergeTreeDataPart::tryGetChecksum(const String & name, const String & ext) const
+{
+    if (checksums.empty())
+        return nullptr;
+
+    const auto & files = checksums.files;
+    const auto file_name = escapeForFileName(name) + ext;
+    auto it = files.find(file_name);
+
+    return (it == files.end()) ? nullptr : &it->second;
+}
+
+const MergeTreeDataPartChecksums::Checksum * MergeTreeDataPart::tryGetBinChecksum(const String & name) const
+{
+    return tryGetChecksum(name, ".bin");
+}
+
+const MergeTreeDataPartChecksums::Checksum * MergeTreeDataPart::tryGetMrkChecksum(const String & name) const
+{
+    return tryGetChecksum(name, ".mrk");
 }
